chore: remove unused types
