--- src/lib/types/global.d.ts
@@ -1,1 +1,0 @@
-export * from "./stats";
