--- src/lib/types/processed/profile/api_settings.d.ts
@@ -1,8 +1,0 @@
-export type APISettings = {
-  skills: boolean;
-  inventory: boolean;
-  personal_vault: boolean;
-  collections: boolean;
-  banking: boolean;
-  museum: boolean;
-};
