--- src/lib/types/processed/profile/enchanting.d.ts
@@ -1,22 +1,0 @@
-export type Enchanting = {
-  unlocked: boolean;
-  data: {
-    [string: string]: ProcessedExperimentationGame;
-  };
-};
-
-export type ProcessedExperimentationGame = {
-  name: string;
-  stats: {
-    lastAttempt: number;
-    lastClaimed: number;
-    bonusClicks: number;
-    games: {
-      name: string;
-      texture: string;
-      attempts: number;
-      claims: number;
-      bestScore: number;
-    }[];
-  };
-};
