--- src/lib/types/processed/profile/rank.d.ts
@@ -1,10 +1,0 @@
-export interface Rank {
-  rankText: string;
-  rankColor: string;
-  plusText: string;
-  plusColor: string;
-}
-
-export type PlusColor = "BLACK" | "DARK_BLUE" | "DARK_GREEN" | "DARK_AQUA" | "DARK_RED" | "DARK_PURPLE" | "GOLD" | "GRAY" | "DARK_GRAY" | "BLUE" | "GREEN" | "AQUA" | "RED" | "LIGHT_PURPLE" | "YELLOW" | "WHITE";
-
-export type RankName = "STAFF" | "OWNER" | "ADMIN" | "GAME_MASTER" | "YOUTUBER" | "SUPERSTAR" | "MVP_PLUS" | "MVP" | "VIP_PLUS" | "VIP" | "MAYOR" | "MINISTER" | "PIG+++";
