--- src/lib/types/processed/profile/rift.d.ts
@@ -1,34 +1,0 @@
-export type Rift = {
-  visits: number;
-  motes: {
-    purse: number;
-    lifetime: number;
-    orbs: number;
-  };
-  enigma: {
-    souls: number;
-    totalSouls: number;
-  };
-  castle: {
-    grubberStacks: number;
-    maxBurgers: number;
-  };
-  porhtal: {
-    porhtalsFound: number;
-    porhtals: {
-      name: string;
-      texture: string;
-      unlocked: boolean;
-    }[];
-  };
-  timecharms: {
-    timecharmsFound: number;
-    timecharms: {
-      name: string;
-      id: string;
-      texture: string;
-      unlocked: boolean;
-      unlockedAt: number;
-    }[];
-  };
-};
