--- src/lib/types/processed/profile/stats.d.ts
@@ -1,16 +1,0 @@
-export type StatsData = {
-  [key: string]: {
-    name: string;
-    nameLore: string;
-    nameShort?: string;
-    nameTiny: string;
-    symbol: string;
-    suffix?: string;
-    color: string;
-    percent?: boolean;
-  };
-};
-
-export type ItemStats = {
-  [key in StatName]?: number;
-};
