--- src/lib/types/stats.d.ts
@@ -1,19 +1,0 @@
-export * from "./processed/profile/api_settings";
-export * from "./processed/profile/bestiary";
-export * from "./processed/profile/collections";
-export * from "./processed/profile/crimson_isle";
-export * from "./processed/profile/dungeons";
-export * from "./processed/profile/enchanting";
-export * from "./processed/profile/farming";
-export * from "./processed/profile/fishing";
-export * from "./processed/profile/items";
-export * from "./processed/profile/main_stats";
-export * from "./processed/profile/mining";
-export * from "./processed/profile/minions";
-export * from "./processed/profile/misc";
-export * from "./processed/profile/pet-stats";
-export * from "./processed/profile/pets";
-export * from "./processed/profile/rank";
-export * from "./processed/profile/rift";
-export * from "./processed/profile/skills";
-export * from "./processed/profile/slayer";
